msautotest: add [ASAN_SKIP_LOG] marker for tests which want to capture stderr
Without the marker if test is run with --run_under_asan the test runner would
append something like '2> asan.log' at the end of the executed command
overriding any stderr redirect from the test itself.
The new marker will suppress this redirect. When used properly errors from
address sanitizer should still be recognizded by those tests because the output
is redirected in the tests result file and will not match the expected output.

+++ msautotest/pymod/mstestlib.py
@@ -651,6 +651,12 @@ def _run(map, out_file, command, extra_args):
 
     (command, strip_items) = collect_strip_requests(command)
 
+    # option to skip asan log output capturing for tests wanting to capture stderr in result file
+    asan_skip_log = False
+    if command.find("[ASAN_SKIP_LOG]") != -1:
+        asan_skip_log = True
+        command = command.replace("[ASAN_SKIP_LOG]", "")
+
     if valgrind:
         valgrind_log = "result/%s.txt" % (out_file + ".vgrind.txt")
         command = command.strip()
@@ -666,7 +672,7 @@ def _run(map, out_file, command, extra_args):
                 + '" | valgrind --tool=memcheck -q --suppressions=../valgrind-suppressions.txt --leak-check=full --show-reachable=yes %s 2>%s'
                 % (command, valgrind_log)
             )
-    elif run_under_asan:
+    elif run_under_asan and not asan_skip_log:
         asan_log = "result/" + out_file + ".asan.txt"
         command = command.strip()
         command += " 2>%s" % asan_log
@@ -701,7 +707,7 @@ def _run(map, out_file, command, extra_args):
             if not quiet:
                 print("     Valgrind log non empty.")
 
-    elif run_under_asan:
+    elif run_under_asan and not asan_skip_log:
         if os.path.getsize(asan_log) == 0:
             os.remove(asan_log)
         else:
